refactor(types): ♻️  organize types
