--- src/types/app.d.ts
@@ -1,90 +1,0 @@
-export type Gender = 'male' | 'female' | 'neutral'
-export type UserRole = 'beta' | 'guest' | 'regular' | 'tester' | 'admin' | 'master'
-
-export interface UserPermissions {
-  activeHuntsLimit: number
-  invitationsLimit: number
-  targetsPerHuntLimit: number
-}
-export interface InterfaceUser {
-  id: string
-  name: string
-  familyName: string
-  accessLevel: string
-  role: UserRole
-  password: string
-  profession?: string
-  gender: Gender
-  birthDate: string
-  email: string
-  lastLogin: string
-  permissions?: UserPermissions
-}
-
-export type SessionUser = Omit<InterfaceUser, 'password'> | undefined
-
-export type SUN_LIGHT = 'morning' | 'afternoon' | 'none'
-
-export interface InterfaceProperty {
-  id: string
-  mainAddressId: string
-  block?: string
-  propertyNumber: string
-  size?: number
-  rooms?: number
-  bathrooms?: number
-  parking?: number
-  is_front?: boolean
-  sun?: SUN_LIGHT
-  condoPricing?: number
-  propertyConvenience: string[]
-}
-
-type PropertyDataProps = Omit<InterfaceProperty, 'id'>
-
-export interface InterfaceLot {
-  id: string
-  name?: string
-  street: string
-  lotNumber: string
-  postalCode?: string
-  neighborhood?: string
-  city: string
-  province: string
-  country: string
-  lotConvenience?: string[]
-}
-
-type LotDataProps = Omit<InterfaceLot, 'id'>
-
-export interface InterfaceAddress extends PropertyDataProps, LotDataProps {}
-
-export type CONTRACT_TYPE = 'buy' | 'rent' | 'either'
-
-export type HuntParticipant = {
-  id: string
-  name: string
-  status: 'waiting' | 'accepted'
-}
-
-export interface InterfaceHunt {
-  id: string
-  creatorId: string
-  huntUsers: HuntParticipant[]
-  title?: string
-  type: CONTRACT_TYPE
-  movingExpected?: string
-  isActive?: boolean
-  livingPeople?: number
-  livingPets?: number
-  targets: string[]
-  minBudget: number
-  maxBudget: number
-}
-
-export type AmenityOf = 'lot' | 'property'
-export interface AmenityData {
-  identifier: string
-  label?: string
-  amenityOf?: AmenityOf
-}
