+++ src/ui/pages/home/components/NextMove/index.tsx
@@ -6,7 +6,7 @@ import cx from 'classnames'
 import { useRouter } from 'next/navigation'
 
 import { getAllTargetPropertiesfromHunt } from '@/requests/targetProperty/getAllTargetProperties'
-import type { CONTRACT_TYPE, InterfaceHunt } from '@/types/app'
+import type { CONTRACT_TYPE, InterfaceHunt } from '@/types/hunt'
 import type { TargetPropertyInterface } from '@/types/targetProperty'
 import Button from '@/ui/components/base/Button'
 import { Loading } from '@/ui/components/base/Loading'
