+++ src/ui/pages/hunt/components/HuntParticipants/index.tsx
@@ -7,7 +7,7 @@ import cx from 'classnames'
 import { useSessionContext } from '@/providers/AuthProvider'
 import { useHuntContext } from '@/providers/HuntProvider'
 import { useUIContext } from '@/providers/UIProvider'
-import type { HuntParticipant, InterfaceHunt } from '@/types/app'
+import type { HuntParticipant, InterfaceHunt } from '@/types/hunt'
 import Button from '@/ui/components/base/Button'
 import { Loading } from '@/ui/components/base/Loading'
 import InviteUserForm from '@/ui/forms/Hunt/InviteUsers'
