+++ src/utils/lang.ts
@@ -1,4 +1,4 @@
-import type { Gender } from '@/types/app'
+import type { Gender } from '@/types/user'
 
 export const genderVowel: { [_key in Gender]: string } = {
   female: 'a',
